creating blueprints to every domain of flask app
