+++ run.py
@@ -1,3 +1,3 @@
 from front_end.web import app
 
-app.run(degub=True)
+app.run(debug=True)
